Move SlotHelpers to sdk/util
