+++ packages/sdk/src/models/item-update.ts
@@ -5,9 +5,10 @@ import {
   PutItemsRequest,
 } from '@meeco/vault-api-sdk';
 import { IDEK } from '../services/service';
+import SlotHelpers from '../util/slot-helpers';
 import { toNestedClassificationNode } from '../util/transformers';
 import ItemChange from './item-change';
-import { NewSlot, SlotHelpers } from './local-slot';
+import { NewSlot } from './local-slot';
 import { NewItem } from './new-item';
 
 /** Construct an update for an existing Item given its id. */
